More test folder fixes.
